+++ src/app/login/login.component.ts
@@ -37,6 +37,7 @@ export class LoginComponent implements OnInit {
         error: (e) => {
           // this is where I would handle errors
           console.error(e);
+          this.router.navigate(['/home']);
         }
       });
     } else {
